Compilation pour integration partielle

+++ RaptorCore/GLHierarchy/Geometry.cpp
@@ -1247,7 +1247,7 @@ bool CGeometry::removePrimitive(CGeometryPrimitive *primitive)
 	}
 }
 
-void CGeometry::glSetPolygons(unsigned int nbP, unsigned short* polygons)
+void CGeometry::glSetPolygons(size_t nbP, unsigned short* polygons)
 {
 	if (polygons == NULL)
 	{
@@ -1276,7 +1276,7 @@ void CGeometry::glSetPolygons(unsigned int nbP, unsigned short* polygons)
 	}
 }
 
-void CGeometry::glSetVertices(unsigned int nbV, GL_COORD_VERTEX* vertices)
+void CGeometry::glSetVertices(size_t nbV, GL_COORD_VERTEX* vertices)
 {
 	if (vertices == NULL)
 	{
@@ -1313,7 +1313,7 @@ void CGeometry::glSetVertices(unsigned int nbV, GL_COORD_VERTEX* vertices)
 #endif
 }
 
-void CGeometry::glSetNormals(unsigned int nbN, GL_COORD_VERTEX* norms)
+void CGeometry::glSetNormals(size_t nbN, GL_COORD_VERTEX* norms)
 {
 #if defined (DATA_PACKED)
 	if ((norms != NULL) && (nbN > 0) && (normals != NULL))
@@ -1327,7 +1327,7 @@ void CGeometry::glSetNormals(unsigned int nbN, GL_COORD_VERTEX* norms)
 #endif
 }
 
-void CGeometry::glSetTexCoords(unsigned int nbT, GL_TEX_VERTEX* texCoords)
+void CGeometry::glSetTexCoords(size_t nbT, GL_TEX_VERTEX* texCoords)
 {
 #if defined (DATA_PACKED)
 	if (texCoords == NULL)
@@ -1351,7 +1351,7 @@ void CGeometry::glSetTexCoords(unsigned int nbT, GL_TEX_VERTEX* texCoords)
 #endif
 }
 
-void CGeometry::glSetTexCoords2(unsigned int nbT, GL_TEX_VERTEX* texCoords)
+void CGeometry::glSetTexCoords2(size_t nbT, GL_TEX_VERTEX* texCoords)
 {
 #if defined (DATA_PACKED)
 	if (texCoords == NULL)
@@ -1375,7 +1375,7 @@ void CGeometry::glSetTexCoords2(unsigned int nbT, GL_TEX_VERTEX* texCoords)
 #endif
 }
 
-void CGeometry::glSetWeights(unsigned int nbW, float* weights)
+void CGeometry::glSetWeights(size_t nbW, float* weights)
 {
 #if defined (DATA_PACKED)
 	if (weights == NULL)
@@ -1399,7 +1399,7 @@ void CGeometry::glSetWeights(unsigned int nbW, float* weights)
 #endif
 }
 
-void CGeometry::glSetColors(unsigned int nbC, CColor::RGBA* rgbaColors)
+void CGeometry::glSetColors(size_t nbC, CColor::RGBA* rgbaColors)
 {
 #if defined (DATA_PACKED)
 	if (rgbaColors == NULL)
@@ -1423,7 +1423,7 @@ void CGeometry::glSetColors(unsigned int nbC, CColor::RGBA* rgbaColors)
 #endif
 }
 
-void CGeometry::glSetFogs(unsigned int nbF, float* fogs)
+void CGeometry::glSetFogs(size_t nbF, float* fogs)
 {
 #if defined (DATA_PACKED)
 	if (fogs == NULL)
@@ -1576,7 +1576,7 @@ void CGeometry::addVertex(float x, float y, float z,float h)
 	}
 }
 
-void CGeometry::setCoord(unsigned int numvtx,GLfloat x,GLfloat y,GLfloat z,GLfloat h)
+void CGeometry::setCoord(size_t numvtx, GLfloat x, GLfloat y, GLfloat z, GLfloat h)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1593,7 +1593,7 @@ void CGeometry::setCoord(unsigned int numvtx,GLfloat x,GLfloat y,GLfloat z,GLflo
 	}
 }
 
-void CGeometry::setTexCoord(unsigned int numvtx,float u,float v)
+void CGeometry::setTexCoord(size_t numvtx, float u, float v)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1638,7 +1638,7 @@ void CGeometry::addFace(int p1,int p2,int p3)
 	}
 }
 
-void CGeometry::getFace(unsigned int numface,unsigned int &p1,unsigned int &p2,unsigned int &p3) const
+void CGeometry::getFace(size_t numface, unsigned int &p1, unsigned int &p2, unsigned int &p3) const
 {
 	if ((numface<m_nbPolys) && (polys != NULL))
 	{
@@ -1652,7 +1652,7 @@ void CGeometry::getFace(unsigned int numface,unsigned int &p1,unsigned int &p2,u
 	}
 }
 
-void CGeometry::setWeight(unsigned int numvtx,float w)
+void CGeometry::setWeight(size_t numvtx, float w)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1662,7 +1662,7 @@ void CGeometry::setWeight(unsigned int numvtx,float w)
 		WEIGHT(numvtx) = w;
 }
 
-float CGeometry::getWeight(unsigned int numvtx) const
+float CGeometry::getWeight(size_t numvtx) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1675,7 +1675,7 @@ float CGeometry::getWeight(unsigned int numvtx) const
 }
 
 
-void CGeometry::setFogCoord(unsigned int numvtx,GLfloat f)
+void CGeometry::setFogCoord(size_t numvtx, GLfloat f)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1685,7 +1685,7 @@ void CGeometry::setFogCoord(unsigned int numvtx,GLfloat f)
 		FOGCOORD(numvtx) = f;
 }
 
-float CGeometry::getFogCoord(unsigned int numvtx) const
+float CGeometry::getFogCoord(size_t numvtx) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1697,7 +1697,7 @@ float CGeometry::getFogCoord(unsigned int numvtx) const
 		return 0.0f;
 }
 
-void CGeometry::getCoord(unsigned int numvtx,GL_COORD_VERTEX &v) const
+void CGeometry::getCoord(size_t numvtx, GL_COORD_VERTEX &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1707,7 +1707,7 @@ void CGeometry::getCoord(unsigned int numvtx,GL_COORD_VERTEX &v) const
 		v = VERTEX(numvtx);
 }
 
-void CGeometry::getVertex(unsigned int numvtx,GL_VERTEX_DATA &v) const
+void CGeometry::getVertex(size_t numvtx, GL_VERTEX_DATA &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1738,7 +1738,7 @@ void CGeometry::getVertex(unsigned int numvtx,GL_VERTEX_DATA &v) const
 #endif
 }
 
-void CGeometry::setNormal(unsigned int numvtx,GLfloat x,GLfloat y,GLfloat z,GLfloat h)
+void CGeometry::setNormal(size_t numvtx, GLfloat x, GLfloat y, GLfloat z, GLfloat h)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1753,7 +1753,7 @@ void CGeometry::setNormal(unsigned int numvtx,GLfloat x,GLfloat y,GLfloat z,GLfl
 	}
 }
 
-void CGeometry::getNormal(unsigned int numvtx,GL_COORD_VERTEX &v) const
+void CGeometry::getNormal(size_t numvtx, GL_COORD_VERTEX &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1763,7 +1763,7 @@ void CGeometry::getNormal(unsigned int numvtx,GL_COORD_VERTEX &v) const
 		v = NORMAL(numvtx);
 }
 
-void CGeometry::getTangent(unsigned int numvtx,GL_COORD_VERTEX &v) const
+void CGeometry::getTangent(size_t numvtx, GL_COORD_VERTEX &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1773,7 +1773,7 @@ void CGeometry::getTangent(unsigned int numvtx,GL_COORD_VERTEX &v) const
 		v = TANGENT(numvtx);
 }
 
-void CGeometry::getBiNormal(unsigned int numvtx,GL_COORD_VERTEX &v) const
+void CGeometry::getBiNormal(size_t numvtx, GL_COORD_VERTEX &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1783,7 +1783,7 @@ void CGeometry::getBiNormal(unsigned int numvtx,GL_COORD_VERTEX &v) const
 		v = BINORMAL(numvtx);
 }
 
-void CGeometry::setColor(unsigned int numvtx,float r,float g,float b,float a)
+void CGeometry::setColor(size_t numvtx, float r, float g, float b, float a)
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
@@ -1798,7 +1798,7 @@ void CGeometry::setColor(unsigned int numvtx,float r,float g,float b,float a)
 	}
 }
 
-void CGeometry::getColor(unsigned int numvtx,CColor::RGBA &v) const
+void CGeometry::getColor(size_t numvtx, CColor::RGBA &v) const
 {
 #if defined (DATA_EXTENDED)
     if ((numvtx<m_nbVertex) && (geometry != NULL))
